fix checking of db.md5, use cmp for comparing
When db.md5 is downloaded and doesn't differ from the existing one, it remains in /tmp.
If radio is then started later again, the next download will create db.md5.1 in /tmp,
at the next start db.md5.2 and so on. But the check wether the file has changed is
always performed against the first downloaded version, db.md5.

Fix that by removing /tmp/db.md5 after comparing.

I also took the oppurtunity to replace  "diff … | wc -l " against "cmp -s".

By the way, if db.md5 is older than a day, the online check occurs with every single start of radio,
not just once a day. Don't know wether this is on purpose or by accident. If the check should be
limited to once a day, we had to "touch" db.md5 after the check unconditionally.